kind of works but could be better
there is currently a tradeoff between accuracy and jitteriness. of course having better accuracy is better than having a nonjittery robot, but it is way to unstable.

the problem is coming from the holonomic drive controller. the higher you make the p value, the more accurate but more jittery it is, and vice versa. maybe could be fixed by meddling with the d?? but the output of the holonomic drive controller is really weird. it is very jumpy. might need to implement our own stabilization function to make it less jumpy